Moved Segment and Coordinate interfaces to Tracks interface
