--- src/interfaces/Coordinate.ts
@@ -1,4 +1,0 @@
-export interface Coordinate {
-    lat: number;
-    lng: number;
-}
